% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{self.traint}
\alias{self.traint}
\title{Co-embedding ST and SC data using Seurat transfer anchors}
\usage{
self.traint(
  st_data,
  sc_data,
  st_assay = "Spatial",
  sc_assay = "RNA",
  norm = "LogNormalize",
  nfeatures = 2000,
  cell_names = "celltype_fine",
  coord_xy = c("imagerow", "imagecol"),
  gene_kept = NULL,
  ...
)
}
\arguments{
\item{st_data}{Seurat ST data object}

\item{sc_data}{Seurat SC data object}

\item{st_assay}{ST assay}

\item{sc_assay}{SC assay}

\item{norm}{normalization method: LogNormalize/SCTransform}

\item{nfeatures}{number of features for integration}

\item{cell_names}{cell cluster/type column name in SC meta data}

\item{coord_xy}{coordinates column names in ST images slot}

\item{gene_kept}{selected genes to be kept during integration}
}
\value{
Seurat object of SC-ST co-embedding
}
\description{
Co-embedding ST and SC data using Seurat transfer anchors
}
\examples{
\donttest{
library(CellTrekSeuratV5)
data(sc_data)
celltrek.int <- self.traint(st_data=st_data, sc_data=sc_data, sc_assay='RNA', cell_names='celltype_fine')
}
}
